added basic permissions to some commands

+++ src/main/java/de/voidtech/gerald/commands/management/SayCommand.java
@@ -1,20 +1,24 @@
-package main.java.de.voidtech.gerald.commands.info;
+package main.java.de.voidtech.gerald.commands.management;
 
 import java.util.List;
 
 import main.java.de.voidtech.gerald.annotations.Command;
 import main.java.de.voidtech.gerald.commands.AbstractCommand;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Message;
 
 @Command
-public class SayCommand extends AbstractCommand{
+public class SayCommand extends AbstractCommand {
 
 	@Override
 	public void executeInternal(Message message, List<String> args) {
-		String msg = String.join(" ", args);
-		message.getChannel().sendMessage(msg).queue(response -> {
-			message.delete().queue();
-		});		
+		if(message.getMember().hasPermission(Permission.MESSAGE_MANAGE))
+		{
+			String msg = String.join(" ", args);
+			message.getChannel().sendMessage(msg).queue(response -> {
+				message.delete().queue();
+			});		
+		}
 	}
 
 	@Override
